+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [AccuWeather] Fix normals (missing in January, shifted by one month on other months)
 
 **Other improvements and fixes**
+- Better precipitation notifications, now based on precipitation minute by minute (not supported by all weather sources)
 - Fix a display issue of “alert source” instead of “normals source” in secondary weather sources settings of a location
 - When background updates are disabled, when opening a location, it will now automatically refresh if it was last refreshed more than 1.5 hours ago
 - When background updates are disabled, make user aware that “Notifications of weather alerts” and “Notifications of precipitations” are unavailable by greying the options in settings